camera: wip

+++ camera/QCamera2/stack/mm-camera-interface/src/mm_camera_interface.c
@@ -41,6 +41,7 @@
 #include <media/msm_cam_sensor.h>
 #include <cutils/properties.h>
 #include <stdlib.h>
+#include <unistd.h>
 
 #include "mm_camera_dbg.h"
 #include "mm_camera_interface.h"
@@ -1542,7 +1543,7 @@ uint8_t get_num_of_cameras()
             break;
         }
 
-        if (strncmp(mdev_info.model, "msm_config", sizeof(mdev_info.model) != 0)) {
+        if (strncmp(mdev_info.model, "msm_config", (size_t)(sizeof(mdev_info.model) != 0))) {
             close(dev_fd);
             dev_fd = 0;
             continue;
